requirements.txt
absl-py==0.15.0
astunparse==1.6.3
beautifulsoup4==4.12.3
bs4==0.0.2
cached-property==1.5.2
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
colorama==0.4.5
#dataclasses==0.8
Flask==2.0.3
flatbuffers==1.12
gast==0.4.0
gdown==4.7.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0 
grpcio==1.48.2
h5py==3.1.0
HTMLParser==0.0.2
idna==3.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.1
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.7
MarkupSafe==2.0.1
nltk==3.6.7
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyMySQL==1.0.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.8.8
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
six==1.15.0
soupsieve==2.3.2.post1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
tqdm==4.64.1
typing-extensions==3.7.4.3
urllib3==1.26.18
Werkzeug==2.0.3
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
